refactor: Extract ScopeAction to get new instance
